Update portfolio website layout and styles
The layout and styles of the portfolio website have been massively updated to display information about work experiences and projects. All the changes include modifying the layout, global CSS and refactoring the code for better readability. Also, a new Header component was added for navigation purposes, and a TimelineItem component for displaying work experiences. Minor icon files were also included in this commit.